Restructure logic
